Reorganized the file structure to follow pages
