--- client/anime-ecomm/src/components/ShopAll.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ShopAll = () => {
-  return <div>SHOP ALL PAGE</div>;
-};
-
-export default ShopAll;
