+++ client/anime-ecomm/src/pages/Clothing.jsx
