+++ client/anime-ecomm/src/pages/Collectables.jsx
