+++ client/anime-ecomm/src/pages/HomeEntertainment.jsx
