+++ client/anime-ecomm/src/pages/HomePage/HomePage.css
