+++ client/anime-ecomm/src/pages/Register.jsx
